Add files via upload
In this lab, I programmed functions in C to take input from a user and validate the input to see if the numbers can form a right triangle.  I then used functions and math.h function to compute the area of this right triangle as well as some of the basic trigonometric function computations. (Sine, Cosine, and Tangent)